BackEnd/requirements.txt
fastapi==0.115.0
pymongo==4.10.1
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.30.6
motor==3.6.0